🐞 fix:修复了聚合函数AVG的INT和FLOAT类型转换的bug

+++ src/observer/sql/expr/aggregator.h
@@ -53,9 +53,9 @@ public:
 class AvgAggregator : public Aggregator
 {
 public:
-  int num = 0;
-  RC  accumulate(const Value &value) override;
-  RC  evaluate(Value &result) override;
+  float num = 0;
+  RC    accumulate(const Value &value) override;
+  RC    evaluate(Value &result) override;
 };
 
 class CountAggregator : public Aggregator
